refactor(layouts): remove index
