--- layouts/container.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-export default styled.div`
-  width: 100%;
-  height: 100%;
-  margin: 0 auto;
-  padding: 1.5rem;
-  max-width: 82.5rem;
-`
