--- layouts/index.ts
@@ -1,7 +1,0 @@
-export {
-  default as TopNav
-} from './topnav'
-
-export {
-  default as Container
-} from './container'
