+++ pages/_app.tsx
@@ -2,9 +2,18 @@ import { Provider } from 'react-redux'
 
 import { ThemeProvider, GlobalStyles } from '@space-metaverse-ag/space-ui'
 import type { AppProps } from 'next/app'
+import styled from 'styled-components'
 
-import { TopNav, Container } from '~/layouts'
+import TopNav from '~/layouts/TopNav'
 import { store } from '~/redux/store'
+
+const Container = styled.div`
+  width: 100%;
+  height: 100%;
+  margin: 0 auto;
+  padding: 1.5rem;
+  max-width: 82.5rem;
+`
 
 const Root: React.FC<AppProps> = ({ Component, pageProps }) => (
   <Provider store={store}>
